Moved State classes into their own files.
