+++ src/LSST/control/StandbyState.h
@@ -19,31 +19,46 @@
  * see <http://www.lsstcorp.org/LegalNotices/>.
  */
 
-// Class header
-#include "control/State.h"
+#ifndef LSST_M2CELLCPP_CONTROL_STANDBYSTATE_H
+#define LSST_M2CELLCPP_CONTROL_STANDBYSTATE_H
+
+// System headers
+#include <functional>
+#include <map>
+#include <memory>
 
 // Project headers
-#include "control/Model.h"
-#include "control/StateMap.h"
-#include "util/Bug.h"
-#include "util/Log.h"
-
-using namespace std;
+#include "control/State.h"
 
 namespace LSST {
 namespace m2cellcpp {
 namespace control {
 
-void State::onEnterState(State::Ptr const& oldState) {
-    LINFO("Entering state=", getName(), " from oldState=", oldState->getName());
-    enterState(oldState);
-}
+class StandbyState : public State {
+public:
+    using Ptr = std::shared_ptr<StandbyState>;
 
-void State::onExitState(State::Ptr const& newState) {
-    LINFO("Leaving state=", getName(), " to go to newState=", newState->getName());
-    exitState(newState);
-}
+    /// Create an instance and insert it into `stateMap`.
+    /// @throws Bug if there's already an instance of this class in `stateMap`.
+    static Ptr create(StateMap& stateMap);
+
+    StandbyState(StandbyState const&) = delete;
+    StandbyState& operator=(StandbyState const&) = delete;
+    virtual ~StandbyState() = default;
+
+    /// VI-PH  exitVI // calls Model::changeStateVI(OfflineState)
+    void exitVI();
+
+    /// VI-PH  startVI // calls Model::startVI  then  Model::stopMotionVI  then
+    /// Model::changeStateVI(ReadyIdle)
+    void startVI();
+
+private:
+    StandbyState() : State("StandbyState") {}
+};
 
 }  // namespace control
 }  // namespace m2cellcpp
 }  // namespace LSST
+
+#endif  // LSST_M2CELLCPP_CONTROL_STANDBYSTATE_H
