+++ src/LSST/control/State.h
@@ -127,177 +127,8 @@ private:
     std::string const _name;  ///< Name of this state.
 };
 
-/// This class represents the state when the system is started.
-/// THis class may be removed in the future if it duplicates the
-/// functionality of another state.
-class StartupState : public State {
-public:
-    using Ptr = std::shared_ptr<StartupState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    StartupState(StartupState const&) = delete;
-    StartupState& operator=(StartupState const&) = delete;
-    virtual ~StartupState() = default;
-
-private:
-    StartupState() : State("StartupState") {}
-};
-
-/// Class representation of the FaultState, this state appears to be unused in LabView!
-class FaultState : public State {
-public:
-    using Ptr = std::shared_ptr<FaultState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    FaultState(FaultState const&) = delete;
-    FaultState& operator=(FaultState const&) = delete;
-    virtual ~FaultState() = default;
-    // VI-PH  clearErrorVI   // calls Model::resetErrorCodeVI then Model::changeStateVI(StandbyState)
-    // VI-PH  goToStandbyVI  // calls Model::resetErrorCodeVI then Model::stopVI then
-    // Model::changeStateVI(StandbyState)
-
-private:
-    FaultState() : State("FaultState") {}
-};
-
-/// Class representation of the IdleState, aka ReadyIdle.
-class IdleState : public State {
-public:
-    using Ptr = std::shared_ptr<IdleState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    IdleState(IdleState const&) = delete;
-    IdleState& operator=(IdleState const&) = delete;
-    virtual ~IdleState() = default;
-
-    /// VI-PH  goToInMotionVI // calls Model::changeStateVI(ReadyInMotion)
-    void goToInMotionVI() override;
-
-    /// VI-PH  goToStandbyVI // calls Model::stopVI then Model::changeStateVI(StandbyState)
-    void goToStandbyVI() override;
-
-    // VI-PH  initScriptEngineVI // calls Model::initScriptEngineVI
-    // VI-PH  loadScriptVI // calls Model::loadScriptVI(scriptFilename)
-    // VI-PH  pauseScriptVI // calls Model::pauseScriptVI
-    // VI-PH  resumeScriptVI // calls Model::resumeScriptVI
-    // VI-PH  setPowerVI // calls Model::setPowerVI(CommPowerControl, MotorPowerControl)   (both power
-    // controls are enums) VI-PH  shutdownCellCommVI // calls Model::shutdownCellComm VI-PH
-    // shutdownMotionEngineVI   // calls Model::shutdownMotionEngine VI-PH  shutdownNetworkInterfaceVI  //
-    // calls Model::shutdownNetworkInterfaceVI VI-PH  shutdownScriptEngineVI // calls
-    // Model::shutdownScriptEngineVI VI-PH  startMotionVI // calls Model::startMotionVI(startMotionParam)
-    // (StartMotionParam includes (Timestamp, MotionStepVector, etc...) VI-PH  startScriptVI // calls
-    // Model::startScriptVI VI-PH stopScriptVI
-    // // call Model::stopScriptVI
-private:
-    IdleState() : State("IdleState") {}
-};
-
-/// Class representation of the InMotionState, aka ReadyInMotion.
-class InMotionState : public State {
-public:
-    using Ptr = std::shared_ptr<InMotionState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    InMotionState(InMotionState const&) = delete;
-    InMotionState& operator=(InMotionState const&) = delete;
-    virtual ~InMotionState() = default;
-
-    /// VI-PH  goToIdleReadyVI // calls Model::changeStateVI(ReadyIdle)
-    void goToIdleReadyVI() override;
-
-    /// VI-PH  goToPauseVI // calls Model::changeStateVI(ReadyPause)
-    void goToPauseVI() override;
-
-    // VI-PH  pauseVI // calls Model::pauseVI
-    // VI-PH  shutdownMotionEngineVI // calls Model::shutdownMotionEngineVI
-    // VI-PH  stopMotionVI // calls Model::stopMotionVI
-    // VI-PH  shutdownMotionEngineVI // calls Model::shutdownMotionEngineVI
-    // VI-PH  stopMotionVI // calls Model::stopMotionVI
-private:
-    InMotionState() : State("InMotionState") {}
-};
-
-/// Class representation of the OfflineState, which leads to program termination.
-class OfflineState : public State {
-public:
-    using Ptr = std::shared_ptr<OfflineState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    OfflineState(OfflineState const&) = delete;
-    OfflineState& operator=(OfflineState const&) = delete;
-    virtual ~OfflineState() = default;
-
-    // nothing here in LabView
-private:
-    OfflineState() : State("OfflineState") {}
-};
-
-/// Class representation of the PauseState, aka ReadyPause.
-class PauseState : public State {
-public:
-    using Ptr = std::shared_ptr<PauseState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    PauseState(PauseState const&) = delete;
-    PauseState& operator=(PauseState const&) = delete;
-    virtual ~PauseState() = default;
-
-    /// VI-PH  goToIdleReadyVI // calls Model::changeStateVI(ReadyIdle)
-    void goToIdleReadyVI() override;
-
-    /// VI-PH  goToInMotionVI // calls Model::changeStateVI(ReadyInMotion)
-    void goToInMotionVI() override;
-
-    // VI-PH  resumeVI // calls Model::resumeMotionVI
-    // VI-PH  shutdownMotionEngineVI // calls Model::shutdownMotionEngineVI
-    // VI-PH  stopMotionVI // calls Model::stopMotionVI
-private:
-    PauseState() : State("PauseState") {}
-};
-
-class StandbyState : public State {
-public:
-    using Ptr = std::shared_ptr<StandbyState>;
-
-    /// Create an instance and insert it into `stateMap`.
-    /// @throws Bug if there's already an instance of this class in `stateMap`.
-    static Ptr create(StateMap& stateMap);
-
-    StandbyState(StandbyState const&) = delete;
-    StandbyState& operator=(StandbyState const&) = delete;
-    virtual ~StandbyState() = default;
-
-    /// VI-PH  exitVI // calls Model::changeStateVI(OfflineState)
-    void exitVI();
-
-    /// VI-PH  startVI // calls Model::startVI  then  Model::stopMotionVI  then
-    /// Model::changeStateVI(ReadyIdle)
-    void startVI();
-
-private:
-    StandbyState() : State("StandbyState") {}
-};
-
 }  // namespace control
 }  // namespace m2cellcpp
 }  // namespace LSST
 
-#endif  // LSST_M2CELLCPP_CONTROL_STATE
+#endif  // LSST_M2CELLCPP_CONTROL_STATE_H
